feat: 对接 Xboard 后端的安全/正确性修复 + 可观测性与可靠性增强
参考 clash-verge-rev 与 FlClash，对全平台客户端做审计与优化。

后端对接 / 认证（修复阻断级缺陷）
- captcha 字段对齐：按 captcha_type 路由到 turnstile_token /
  recaptcha_v3_token / recaptcha_data（修 Turnstile/v3 面板下注册恒 400）
- 订阅 4xx 全链路：client/fetcher/manager 对 403 等不写盘、不污染缓存、
  上抛专用错误，过期用户不再被当作"已连接"；node 预览返回续费引导
- API 层与后端路由逐项核对修正

连接可用性
- 内核崩溃(Exited)/长时间无响应自动清系统代理 + 回收 handle + 置 Error
- 系统代理守卫：30s 巡检，被其它程序篡改后自动夺回（可在设置页关闭）
- 自动重连：崩溃后按 1s/3s/6s 退避自动恢复
- 动态空闲端口（controller/mixed），根除与其它 Clash 软件的端口冲突
- 切节点后 flush 连接（不残留脏连接）；断开/退出逐项超时，永不卡死
- connect 跨 await unwrap 改为安全分支，消除登出竞态 panic

安全（提权辅助进程）
- helper(macOS) 经 SO_PEERCRED 校验连接方 uid，cfg/work/log 路径锁死在
  该用户自有 Application Support 子目录；svc(Windows) 镜像路径白名单
- 修正 IPC 模块误导性的"已实现 HMAC"注释为真实边界描述

移动端数据面（core 侧打通；平台侧需真机验证）
- Android：ffi 连接路径调用 TunDelegate.establish_tun 并把 fd 注入
  mihomo（file-descriptor + 关 auto-route），修"从不注入 fd"
- iOS：PacketTunnelProvider 改为返回真实 utun fd（wireguard-apple 手法）
- sing-box 生成 route.rule_set[] 定义，修 1.8+ "rule-set not found" 启动失败

可观测性套件（桌面）
- 实时连接监控页、实时日志页、分流规则页、上下行流量折线图
- core 暴露 connections/close/rules；日志流经 connection://log 转发

桌面便利
- 厚托盘：连接/断开/模式/日志菜单 + 状态实时 tooltip
- 集中设置页（主题/语言/默认模式/代理守卫/启动即连/版本）
- 节点选择持久化（重启恢复上次手选节点）；老用户启动即连

文档
- 审计报告.md、优化方案-参考FlClash与clash-verge.md

验证：cargo check --workspace + clippy 零告警 + core 68 测试 + vue-tsc + vite build 全绿

+++ core/src/api/user.rs
@@ -54,8 +54,14 @@ pub struct SubscribeInfo {
     /// Full subscription URL — opaque to the client; pass it to the kernel
     /// after appending the right `?flag=` for the active driver.
     pub subscribe_url: String,
+    /// Days remaining until the next traffic reset. For monthly-reset plans
+    /// this fits comfortably in a `u8`, but the yearly-reset variants
+    /// (`RESET_TRAFFIC_YEARLY` / `RESET_TRAFFIC_FIRST_DAY_YEAR`) legitimately
+    /// emit values up to ~366, which overflows `u8` and causes the whole
+    /// `/user/getSubscribe` response to fail to deserialize on yearly-reset
+    /// deployments. `u16` covers any plausible reset horizon.
     #[serde(default)]
-    pub reset_day: Option<u8>,
+    pub reset_day: Option<u16>,
 }
 
 impl HttpClient {
